# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: nGX API
  version: 1.0.0
  description: |
    Email infrastructure purpose-built for AI agents.

    nGX lets you provision inboxes, send and receive email, manage
    drafts with human-in-the-loop review, subscribe to real-time events via
    webhooks or WebSockets, and search across your entire message history.

    ## Authentication
    All API requests (except `/health`) require a Bearer API key in the
    `Authorization` header:
    ```
    Authorization: Bearer am_live_...
    ```
    Keys are scoped — see the `scopes` field on the key endpoints.

    ## Organization And Keys
    Organizations are created during account signup and are discovered at runtime
    via the authenticated API key. `GET /org` returns the current organization
    for that key.

    `POST /keys` creates additional API keys for the current organization.
    It does not create a new organization or perform account registration.
  contact:
    name: nGX Support
    url: https://yourdomain.com
servers:
  - url: https://api.yourdomain.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Tags
# ---------------------------------------------------------------------------
...
paths:
...
  /pods:
    get:
      tags: [Pods]
      operationId: listPods
      summary: List pods
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/pageToken"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PodList"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Pods]
      operationId: createPod
      summary: Create a pod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePodRequest"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pod"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /pods/{podId}:
    parameters:
      - $ref: "#/components/parameters/podId"
    get:
      tags: [Pods]
      operationId: getPod
      summary: Get a pod
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pod"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      tags: [Pods]
      operationId: updatePod
      summary: Update a pod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePodRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pod"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Pods]
      operationId: deletePod
      summary: Delete a pod
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatus"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  # -------------------------------------------------------------------------
  # Inboxes
  # -------------------------------------------------------------------------
...
  /inboxes/{inboxId}/threads/{threadId}/messages:
    parameters:
      - $ref: "#/components/parameters/inboxId"
      - $ref: "#/components/parameters/threadId"
    get:
      tags: [Messages]
      operationId: listMessages
      summary: List messages in a thread
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/pageToken"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /inboxes/{inboxId}/threads/{threadId}/messages/{messageId}:
    parameters:
      - $ref: "#/components/parameters/inboxId"
      - $ref: "#/components/parameters/threadId"
      - $ref: "#/components/parameters/messageId"
    get:
      tags: [Messages]
      operationId: getMessage
      summary: Get a message
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      tags: [Messages]
      operationId: updateMessage
      summary: Update message metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMessageRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /inboxes/{inboxId}/messages/send:
    parameters:
      - $ref: "#/components/parameters/inboxId"
    post:
      tags: [Messages]
      operationId: sendMessage
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  # -------------------------------------------------------------------------
  # Drafts (nested under inboxes)
  # -------------------------------------------------------------------------
...
  /webhooks:
    get:
      tags: [Webhooks]
      operationId: listWebhooks
      summary: List webhooks
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/pageToken"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookList"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Webhooks]
      operationId: createWebhook
      summary: Create a webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /webhooks/{webhookId}:
    parameters:
      - $ref: "#/components/parameters/webhookId"
    get:
      tags: [Webhooks]
      operationId: getWebhook
      summary: Get a webhook
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      tags: [Webhooks]
      operationId: updateWebhook
      summary: Update a webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWebhookRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Webhooks]
      operationId: deleteWebhook
      summary: Delete a webhook
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatus"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "API Key (am_live_...)"
      description: |
        All API requests require a Bearer API key.
        Keys can be scoped to specific capabilities and pods.

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
...
    podId:
      name: podId
      in: path
      required: true
      description: Pod ID.
      schema:
        type: string
    inboxId:
      name: inboxId
      in: path
      required: true
      description: Inbox ID.
      schema:
        type: string
    threadId:
      name: threadId
      in: path
      required: true
      description: Thread ID.
      schema:
        type: string
    messageId:
      name: messageId
      in: path
      required: true
      description: Message ID.
      schema:
        type: string
...
    webhookId:
      name: webhookId
      in: path
      required: true
      description: Webhook ID.
      schema:
        type: string
...
    limit:
      name: limit
      in: query
      description: Maximum number of items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    pageToken:
      name: page_token
      in: query
      description: Opaque page token from a previous response.
      schema:
        type: string

  # -------------------------------------------------------------------------
  # Responses
  # -------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Bad Request — invalid parameters or request body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "validation failed"
            details:
              field: "url"
              reason: "must be a valid https URL"
    Unauthorized:
      description: Unauthorized — missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "missing or invalid authorization header"
...
    NotFound:
      description: Not Found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "not found"
    Conflict:
      description: Conflict — resource already exists or state transition is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "email address already registered"
...
    InternalError:
      description: Internal Server Error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "internal server error"

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    OperationStatus:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        resource:
          type: string
        id:
          type: string
          format: uuid
      example:
        success: true
        message: "resource deleted"
        resource: "inbox"
        id: "3f2d9d77-89b9-4b8f-9f47-d0b78664b6cb"
...
    Pod:
      type: object
      required: [pod_id, name, client_id, created_at, updated_at]
      properties:
        pod_id:
          type: string
        name:
          type: string
          example: "Customer Support"
        client_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        pod_id: "pod_123"
        name: "Customer Support"
        client_id: "client_123"
        created_at: "2026-03-27T18:31:00Z"
        updated_at: "2026-03-27T18:31:00Z"
    PodList:
      type: object
      required: [count, pods]
      properties:
        count:
          type: integer
        pods:
          type: array
          items:
            $ref: "#/components/schemas/Pod"
        limit:
          type: integer
          nullable: true
        next_page_token:
          type: string
          nullable: true
      example:
        count: 1
        pods:
          - pod_id: "pod_123"
            name: "Customer Support"
            client_id: "client_123"
            created_at: "2026-03-27T18:31:00Z"
            updated_at: "2026-03-27T18:31:00Z"
        limit: 20
        next_page_token: null
    CreatePodRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "Customer Support"
        slug:
          type: string
          minLength: 1
          maxLength: 63
          pattern: "^[a-z0-9-]+$"
          example: "customer-support"
        description:
          type: string
          maxLength: 1000
        settings:
          type: object
          additionalProperties: true
    UpdatePodRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        settings:
          type: object
          additionalProperties: true

    # --- Inbox ----------------------------------------------------------------
...
    Message:
      type: object
      required: [message_id, inbox_id, thread_id, created_at, updated_at]
      properties:
        message_id:
          type: string
        inbox_id:
          type: string
        thread_id:
          type: string
        subject:
          type: string
        from:
          type: string
          description: RFC 5322 from header rendered as a string.
        to:
          type: array
          items:
            type: string
        cc:
          type: array
          items:
            type: string
        bcc:
          type: array
          items:
            type: string
        reply_to:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        labels:
          type: array
          items:
            type: string
        text:
          type: string
        html:
          type: string
        preview:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        message_id: "msg_123"
        inbox_id: "support@example.com"
        thread_id: "thread_123"
        subject: "Re: Invoice #1234"
        from: "Acme Support <support@example.com>"
        to:
          - "Vendor Billing <billing@vendor.com>"
        cc: []
        bcc: []
        reply_to: []
        labels:
          - "urgent"
        text: "Thanks, we have processed the invoice."
        html: "<p>Thanks, we have processed the invoice.</p>"
        preview: "Thanks, we have processed the invoice."
        headers:
          Message-ID: "<msg_123>"
        attachments: []
        created_at: "2026-03-27T18:39:00Z"
        updated_at: "2026-03-27T18:40:00Z"
    MessageList:
      type: object
      required: [count, messages]
      properties:
        count:
          type: integer
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        limit:
          type: integer
          nullable: true
        next_page_token:
          type: string
          nullable: true
      example:
        count: 1
        messages:
          - message_id: "msg_123"
            inbox_id: "support@example.com"
            thread_id: "thread_123"
            subject: "Re: Invoice #1234"
            from: "Acme Support <support@example.com>"
            to:
              - "Vendor Billing <billing@vendor.com>"
            cc: []
            bcc: []
            reply_to: []
            labels:
              - "urgent"
            text: "Thanks, we have processed the invoice."
            html: "<p>Thanks, we have processed the invoice.</p>"
            preview: "Thanks, we have processed the invoice."
            headers:
              Message-ID: "<msg_123>"
            attachments: []
            created_at: "2026-03-27T18:39:00Z"
            updated_at: "2026-03-27T18:40:00Z"
        limit: 20
        next_page_token: null
    SendMessageRequest:
      type: object
      required: [to]
      properties:
        thread_id:
          type: string
          description: If set, this message is associated with an existing thread.
        in_reply_to:
          type: string
        subject:
          type: string
          example: "Re: Invoice #1234"
        to:
          type: array
          minItems: 1
          items:
            type: string
        cc:
          type: array
          items:
            type: string
        bcc:
          type: array
          items:
            type: string
        reply_to:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        text:
          type: string
        html:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentUpload"
        headers:
          type: object
          additionalProperties:
            type: string
    SendMessageResponse:
      type: object
      required: [message_id, thread_id]
      properties:
        message_id:
          type: string
        thread_id:
          type: string
      example:
        message_id: "msg_123"
        thread_id: "thread_123"
    UpdateMessageRequest:
      type: object
      properties:
        add_labels:
          type: array
          items:
            type: string
        remove_labels:
          type: array
          items:
            type: string
        unread:
          type: boolean
        starred:
          type: boolean
...
    Webhook:
      type: object
      required: [webhook_id, url, event_types, enabled, created_at, updated_at]
      properties:
        webhook_id:
          type: string
        url:
          type: string
          format: uri
        description:
          type: string
        event_types:
          type: array
          items:
            type: string
        pod_ids:
          type: array
          items:
            type: string
        inbox_ids:
          type: array
          items:
            type: string
        enabled:
          type: boolean
        auth_header:
          allOf:
            - $ref: "#/components/schemas/WebhookAuthHeader"
          nullable: true
          description: |
            Present (with only `name`) when a caller-supplied auth header is
            configured. Absent when none is set.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        webhook_id: "wh_123"
        url: "https://example.com/webhooks/nGX"
        description: "Production agent automation"
        event_types:
          - "message.received"
          - "draft.created"
        pod_ids:
          - "pod_123"
        inbox_ids: []
        enabled: true
        created_at: "2026-03-27T18:15:00Z"
        updated_at: "2026-03-27T18:20:00Z"
    WebhookList:
      type: object
      required: [count, webhooks]
      properties:
        count:
          type: integer
        webhooks:
          type: array
          items:
            $ref: "#/components/schemas/Webhook"
        limit:
          type: integer
          nullable: true
        next_page_token:
          type: string
          nullable: true
      example:
        count: 1
        webhooks:
          - webhook_id: "wh_123"
            url: "https://example.com/webhooks/nGX"
            description: "Production agent automation"
            event_types:
              - "message.received"
            pod_ids:
              - "pod_123"
            inbox_ids: []
            enabled: true
            created_at: "2026-03-27T18:15:00Z"
            updated_at: "2026-03-27T18:20:00Z"
        limit: 20
        next_page_token: null
    CreateWebhookRequest:
      type: object
      required: [url, event_types]
      properties:
        url:
          type: string
          format: uri
        description:
          type: string
        event_types:
          type: array
          minItems: 1
          items:
            type: string
        pod_ids:
          type: array
          items:
            type: string
        inbox_ids:
          type: array
          items:
            type: string
        auth_header:
          $ref: "#/components/schemas/WebhookAuthHeader"
          description: |
            Optional. Provide both `name` and `value`. nGX encrypts the
            value at rest and injects the header on every delivery call.
            The `secret` returned on creation is separate — it is used for
            HMAC payload signing so the receiver can verify the call came from
            nGX.
...